refactor : toString exclude 및 패키지 위치 변경
