+++ src/main/java/com/doongji/nestalk/entity/user/User.java
@@ -1,9 +1,9 @@
 package com.doongji.nestalk.entity.user;
 
 import com.doongji.nestalk.entity.BaseTimeEntity;
+import com.doongji.nestalk.entity.friend.Friend;
 import com.doongji.nestalk.security.Jwt;
 import lombok.*;
-import org.hibernate.annotations.JoinColumnOrFormula;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import javax.persistence.*;
@@ -20,7 +20,7 @@ import static org.apache.commons.lang3.StringUtils.isNotEmpty;
 @Entity
 @EqualsAndHashCode(of = "userId", callSuper = false)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@ToString
+@ToString(exclude = {"friends"})
 public class User extends BaseTimeEntity {
 
     @Id
@@ -42,7 +42,7 @@ public class User extends BaseTimeEntity {
     @Column(nullable = false)
     private LocalDate birthday;
 
-    @OneToMany(mappedBy = "me", fetch = FetchType.LAZY)
+    @OneToMany(mappedBy = "me")
     private List<Friend> friends = new ArrayList<>();
 
     public User(String email, String name, String password, String phone, LocalDate birthday) {
