+++ src/main/java/com/doongji/nestalk/repository/friend/FriendRepository.java
@@ -1,9 +1,7 @@
 package com.doongji.nestalk.repository.friend;
 
-import com.doongji.nestalk.entity.user.Friend;
+import com.doongji.nestalk.entity.friend.Friend;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 
 public interface FriendRepository extends JpaRepository<Friend, Long> {
